修改立即关机的bug,增加操作前判断

+++ src/com/service/PCService.java
@@ -5,7 +5,7 @@ import com.util.TurnOffPC;
 public class PCService {
 	private TurnOffPC turnOff = new TurnOffPC();
 	
-	public void turnOffPC(int time){
+	public void turnOffPC(double time){
 		turnOff.turnOffPC(time);
 	}
 	public void cancelOffPC(){
